1055 - expires_in is supposed to be Integer, so RefreshTokenResponse now accepts Integer instead of Long

+++ server/ee/libs/platform/platform-connection/platform-connection-remote-client/src/main/java/com/bytechef/ee/platform/connection/remote/client/fasade/RemoteConnectionFacadeClient.java
@@ -33,7 +33,7 @@ public class RemoteConnectionFacadeClient implements ConnectionFacade {
     }
 
     @Override
-    public Long executeConnectionRefresh(Long connectionId) {
+    public Integer executeConnectionRefresh(Long connectionId) {
         throw new UnsupportedOperationException();
     }
